# docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "Skill Assessment Service:",
    "version": "1.0.0"
  },
  "host": "localhost:8084",
  "basePath": "/",
  "paths": {
    "/api/categories": {
      "get": {
        "security": [
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Fetch all categories",
        "tags": [
          "category"
        ],
        "operationId": "listCategories",
        "responses": {
          "200": {
            "$ref": "#/responses/CategoriesResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/api/category": {
      "post": {
        "security": [
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Add new category",
        "tags": [
          "category"
        ],
        "operationId": "CategoryRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CategoryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/SuccessResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestErrorResponse"
          },
          "409": {
            "$ref": "#/responses/ConflictErrorResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      }
    },
...
    "/api/question/types": {
      "get": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          },
          {
            "securityDefinitions": []
          },
          {
            "Bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Get all question types",
        "tags": [
          "questions"
        ],
        "operationId": "listQuestionTypes",
        "responses": {
          "200": {
            "$ref": "#/responses/QuestionTypeResponse"
          },
          "401": {
            "$ref": "#/responses/UnAuthorizedAccessResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/api/token/refresh": {
      "post": {
        "description": "requests:",
        "tags": [
          "auth"
        ],
        "summary": "Refresh token with existing token",
        "operationId": "RefreshTokenRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/LoginResponse"
          },
          "400": {
            "$ref": "#/responses/InvalidTokenResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/api/user": {
      "post": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          },
          {
            "securityDefinitions": []
          },
          {
            "Bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          },
          {
            "requests": []
          }
        ],
        "description": "Create new user",
        "tags": [
          "admin"
        ],
        "operationId": "UserRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/SuccessResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestErrorResponse"
          },
          "401": {
            "$ref": "#/responses/UnAuthenticatedAccessResponse"
          },
          "403": {
            "$ref": "#/responses/UnAuthorizedAccessResponse"
          },
          "409": {
            "$ref": "#/responses/ConflictErrorResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/api/user/login": {
      "post": {
        "description": "requests:",
        "tags": [
          "auth"
        ],
        "summary": "Login with username and password",
        "operationId": "LoginRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/LoginResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestErrorResponse"
          },
          "401": {
            "$ref": "#/responses/UnAuthorizedResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/api/user/password/reset": {
      "put": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          },
          {
            "securityDefinitions": []
          },
          {
            "Bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Reset user password",
        "tags": [
          "user"
        ],
        "operationId": "ResetUserPasswordRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/PasswordReset"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/LoginResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestErrorResponse"
          },
          "401": {
            "$ref": "#/responses/UnAuthorizedAccessResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundErrorResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/api/user/roles": {
      "get": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          },
          {
            "securityDefinitions": []
          },
          {
            "Bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Get all roles",
        "tags": [
          "admin"
        ],
        "operationId": "listRoles",
        "responses": {
          "200": {
            "$ref": "#/responses/RolesResponse"
          },
          "401": {
            "$ref": "#/responses/UnAuthorizedAccessResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/api/user/{Username}": {
      "delete": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          },
          {
            "securityDefinitions": []
          },
          {
            "Bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          },
          {
            "requests": []
          }
        ],
        "description": "Delete user",
        "tags": [
          "admin"
        ],
        "operationId": "DeleteUserRequest",
        "parameters": [
          {
            "type": "string",
            "example": "username",
            "name": "Username",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/UserDeleteSuccessResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestErrorResponse"
          },
          "401": {
            "$ref": "#/responses/UnAuthorizedAccessResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundErrorResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/api/user/{Username}/password": {
      "put": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          },
          {
            "securityDefinitions": []
          },
          {
            "Bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Reset user password- only for admin",
        "tags": [
          "admin"
        ],
        "operationId": "ResetPasswordRequest",
        "parameters": [
          {
            "type": "string",
            "example": "admin",
            "name": "Username",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/PasswordReset"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResetUserPasswordResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestErrorResponse"
          },
          "401": {
            "$ref": "#/responses/UnAuthorizedAccessResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundErrorResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/api/user/{Username}/revoke": {
      "put": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          },
          {
            "securityDefinitions": []
          },
          {
            "Bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          },
          {
            "requests": []
          }
        ],
        "description": "Revoke user",
        "tags": [
          "admin"
        ],
        "operationId": "RevokeUserRequest",
        "parameters": [
          {
            "type": "string",
            "example": "username",
            "name": "Username",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/UserRevokeSuccessResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestErrorResponse"
          },
          "401": {
            "$ref": "#/responses/UnAuthorizedAccessResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundErrorResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          },
          {
            "securityDefinitions": []
          },
          {
            "Bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Get all available users",
        "tags": [
          "admin"
        ],
        "operationId": "GetAllUsersRequest",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "example": 1,
            "x-go-name": "Page",
            "description": "Provide page number",
            "name": "page",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "example": 10,
            "x-go-name": "PageSize",
            "description": "Provide page size",
            "name": "pageSize",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "example": "admin",
            "x-go-name": "Search",
            "description": "Provide search literal",
            "name": "search",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "example": "username",
            "x-go-name": "SortBy",
            "description": "Provide sort by field (username|email|createdAt|updatedAt)",
            "name": "sortBy",
            "in": "query"
          },
          {
            "type": "string",
            "example": "asc",
            "x-go-name": "OrderBy",
            "description": "provide order by asc|desc",
            "name": "orderBy",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/UsersResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestErrorResponse"
          },
          "401": {
            "$ref": "#/responses/UnAuthorizedAccessResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundEmptyErrorResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/api/{Category}/question": {
      "post": {
        "security": [
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Add new question based on category",
        "tags": [
          "questions"
        ],
        "operationId": "QuestionRequest",
        "parameters": [
          {
            "type": "string",
            "example": "go",
            "name": "Category",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/QuestionRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/SuccessResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestErrorResponse"
          },
          "409": {
            "$ref": "#/responses/ConflictErrorResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      }
    },
    "/api/{Category}/questions": {
      "get": {
        "security": [
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Add new question based on category",
        "tags": [
          "questions"
        ],
        "operationId": "GetQuestionRequest",
        "parameters": [
          {
            "type": "string",
            "example": "go",
            "name": "Category",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/QuestionsResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestErrorResponse"
          },
          "409": {
            "$ref": "#/responses/ConflictErrorResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      },
      "post": {
        "security": [
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Fetch all categories",
        "tags": [
          "questions"
        ],
        "operationId": "QuestionsRequest",
        "parameters": [
          {
            "type": "string",
            "example": "go",
            "name": "Category",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/QuestionRequest"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/SuccessResponse"
          },
          "400": {
            "$ref": "#/responses/BadRequestErrorResponse"
          },
          "500": {
            "$ref": "#/responses/InternalServerErrorResponse"
          }
        }
      }
    }
  },
  "definitions": {
...
    "CategoryRequest": {
      "description": "Category Request",
      "type": "object",
      "required": [
        "categoryName",
        "collectionName",
        "author"
      ],
      "properties": {
        "author": {
          "description": "Author of this category",
          "type": "string",
          "minLength": 5,
          "x-go-name": "Author",
          "example": "admin"
        },
        "categoryName": {
          "description": "the category name",
          "type": "string",
          "minLength": 2,
          "x-go-name": "CategoryName",
          "example": "go"
        },
        "collectionName": {
          "description": "the collection name",
          "type": "string",
          "minLength": 2,
          "x-go-name": "CollectionName",
          "example": "go"
        }
      },
      "x-go-package": "github.com/sandy0786/skill-assessment-service/request/category"
    },
...
  },
  "responses": {
    "BadRequestErrorResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/BadRequestError"
      }
    },
    "CategoriesResponse": {
      "description": "",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/CategoryResponse"
        }
      }
    },
    "ConflictErrorResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/ConflictError"
      }
    },
...
    "InternalServerErrorResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/GlobalError"
      }
    },
...
    "LoginResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/LoginResponse"
      }
    },
...
    "QuestionTypeResponse": {
      "description": "List of questions types",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/QuestionTypeResponse"
        }
      }
    },
...
    "SuccessResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/SuccessResponse"
      }
    },
...
    "UnAuthorizedAccessResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/UnAuthorizedAccessError"
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "Bearer": [
        "[]"
      ]
    }
  ]
}